Fix for 409: assert on mask target must have NoData defined
